Pass events instead of ViewModel to OnBoardingViews
No need to have a separate DateScreenContent
